s10soc: Add new feature for ad_cpu_interconnect
If we have a lot of peripherals connected to the CPU's memory interface,
the generated interconnect can grow to much decreasing the timing
margin.

One solution is to group the peripherals by its interface types and
functions and use bridges to connect them to the memory interface.

This commit adds the possibility to insert an Avalon Memory Mapped
bridge when we create the connection between the peripheral and CPU.
Should be used just with Avalaon Memory Mapped interfaces.